Fixed UV readings bug, added abs() on some display values

+++ Atollic/Diplomski_Outdoor_Unit/myDrivers/Inc/myStructs.h
@@ -11,7 +11,7 @@ struct syncStructHandle
 
 struct measruementHandle
 {
-    uint16_t uv;
+    float uv;
     int16_t windDir;
     float tempSHT;
     float tempSoil;
